+++ packages/web-app/src/components/wrappers/index.ts
@@ -1,0 +1,2 @@
+export * from './pageWrappers';
+export * from './sectionWrappers';
